Update Manifest
Currently the distribution is uninstallable because lib/Config/TOML/Parser.pm
was not listed in the manifest. This commit updates the manifest to include
this file and the other unlisted tests.

This commmit also renames ignore.txt to MANIFEST.SKIP and adds .gitignore
and .git/* to it to satisfy t/manifest.t.